Move ZimContentProvider to data package
