+++ app/src/main/java/org/kiwix/kiwixmobile/KiwixErrorActivity.java
@@ -11,6 +11,7 @@ import android.widget.Button;
 import android.widget.CheckBox;
 
 import org.kiwix.kiwixmobile.base.BaseActivity;
+import org.kiwix.kiwixmobile.data.ZimContentProvider;
 import org.kiwix.kiwixmobile.data.local.dao.BookDao;
 import org.kiwix.kiwixmobile.library.entity.LibraryNetworkEntity;
 import org.kiwix.kiwixmobile.utils.SplashActivity;
