+++ app/src/main/java/org/kiwix/kiwixmobile/data/local/dao/RecentSearchDao.java
@@ -20,7 +20,7 @@ package org.kiwix.kiwixmobile.data.local.dao;
 import com.yahoo.squidb.data.SquidCursor;
 import com.yahoo.squidb.sql.Query;
 
-import org.kiwix.kiwixmobile.ZimContentProvider;
+import org.kiwix.kiwixmobile.data.ZimContentProvider;
 import org.kiwix.kiwixmobile.data.local.KiwixDatabase;
 import org.kiwix.kiwixmobile.data.local.entity.RecentSearch;
